update admin role permission

+++ src/modules/permissions/permissions.service.ts
@@ -1,7 +1,14 @@
-import { HttpException, HttpStatus, Inject, Injectable, Logger, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpException,
+  HttpStatus,
+  Inject,
+  Injectable,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreatePermissionDto } from './dto/create-permission.dto';
 import { UpdatePermissionDto } from './dto/update-permission.dto';
-import { roles } from '../drizzle/schema/schema';
 import { PageDto, PageMetaDto, PageOptionsDto } from '@/core/dto';
 import { DrizzleDB } from '../drizzle/types/drizzle';
 import { DRIZZLE } from '../drizzle/drizzle.module';
@@ -9,6 +16,7 @@ import { PgColumn } from 'drizzle-orm/pg-core';
 import { adminPermission, permissionRole, permissions } from '../drizzle/schema/admin-module.schema';
 import { and, asc, desc, eq, like, not, or, sql } from 'drizzle-orm';
 import slugify from 'slugify';
+import { PermissionModel } from './entities/permission.entity';
 
 @Injectable()
 export class PermissionsService {
@@ -16,7 +24,7 @@ export class PermissionsService {
   constructor(@Inject(DRIZZLE) private readonly db: DrizzleDB) {}
 
   // get all admins
-  async findAll(query: PageOptionsDto) {
+  async findAll(query: PageOptionsDto): Promise<PageDto<PermissionModel>> {
     const limit: number = query.limit || 10;
     const page: number = query.page || 1;
     const search = query.search || '';
@@ -28,12 +36,47 @@ export class PermissionsService {
     }
 
     const order = query.order || 'asc';
-    const isActive = query.isActive || undefined;
+    let isDeleted: boolean = false;
+    let isActive: boolean | undefined;
+
+    switch (query.isActive) {
+      case 'true':
+        isActive = true;
+        break;
+      case 'false':
+        isActive = false;
+        break;
+      default:
+        isActive = true;
+        break;
+    }
+
+    switch (query.isDeleted) {
+      case 'true':
+        isDeleted = true;
+        break;
+      case 'false':
+        isDeleted = false;
+        break;
+      default:
+        isDeleted = false;
+        break;
+    }
 
     const dbQuery = this.db
       .select()
       .from(permissions)
-      .where(and(or(like(permissions.name, `%${search}%`))))
+      .where(
+        and(
+          or(
+            like(permissions.name, `%${search}%`),
+            like(permissions.slug, `%${search}%`),
+            like(permissions.group, `%${search}%`),
+          ),
+          isDeleted ? eq(permissions.isDeleted, isDeleted) : undefined,
+          isActive ? eq(permissions.isActive, isActive) : undefined,
+        ),
+      )
       .$dynamic();
 
     const total = await this.db.$count(dbQuery);
@@ -42,7 +85,21 @@ export class PermissionsService {
 
     const resultQuery = await this.db.query.permissions.findMany({
       orderBy: order === 'desc' ? [desc(sort)] : [asc(sort)],
-      where: and(or(like(permissions.name, `%${search}%`))),
+      where: and(
+        or(
+          like(permissions.name, `%${search}%`),
+          like(permissions.slug, `%${search}%`),
+          like(permissions.group, `%${search}%`),
+        ),
+        isDeleted ? eq(permissions.isDeleted, isDeleted) : undefined,
+        isActive ? eq(permissions.isActive, isActive) : undefined,
+      ),
+      columns: {
+        isDeleted: false,
+        deletedAt: false,
+        deletedBy: false,
+        deletedReason: false,
+      },
       extras: {
         adminCount: sql<number>`(
           SELECT COUNT(*)::int 
@@ -56,7 +113,11 @@ export class PermissionsService {
         )`.as('roleCount'),
       },
       with: {
-        roles: true,
+        roles: {
+          with: {
+            role: true,
+          },
+        },
       },
       limit: limit,
       offset: (page - 1) * limit,
@@ -78,7 +139,7 @@ export class PermissionsService {
 
   async findOne(id: number) {
     const item = await this.db.query.permissions.findFirst({
-      where: eq(permissions.id, Number(id)),
+      where: and(eq(permissions.id, Number(id))),
       extras: {
         adminCount: sql<number>`(
           SELECT COUNT(*)::int 
@@ -100,6 +161,10 @@ export class PermissionsService {
       throw new NotFoundException('Permission not found');
     }
 
+    if (item.isDeleted) {
+      throw new BadRequestException('Permission is deleted');
+    }
+
     return item;
   }
 
@@ -111,17 +176,24 @@ export class PermissionsService {
     });
 
     if (checkItem) {
-      throw new HttpException('Role already exists ', HttpStatus.BAD_REQUEST);
+      throw new HttpException('Permission already exists ', HttpStatus.BAD_REQUEST);
     }
 
-    const insertedRole = await this.db
+    const insertedPermission = await this.db
       .insert(permissions)
       .values({
         name: createDto.name,
         slug: slug,
         group: createDto.group,
+        groupOrder: createDto.groupOrder,
+        order: createDto.order,
+        isActive: createDto.isActive,
         createdAt: new Date(),
         updatedAt: new Date(),
+        isDeleted: false,
+        deletedAt: null,
+        deletedBy: null,
+        deletedReason: null,
       })
       .returning();
 
@@ -133,14 +205,42 @@ export class PermissionsService {
   // get admin by id
   async findById(id: number) {
     const item = await this.db.query.permissions.findFirst({
-      where: eq(permissions.id, id),
+      where: and(eq(permissions.id, Number(id)), eq(permissions.isDeleted, false)),
+      extras: {
+        adminCount: sql<number>`(
+          SELECT COUNT(*)::int 
+          FROM ${adminPermission} ap
+          WHERE ap.permission_id = ${permissions.id}
+        )`.as('adminCount'),
+        roleCount: sql<number>`(
+          SELECT COUNT(*)::int 
+          FROM ${permissionRole} pr
+          WHERE pr.permission_id = ${permissions.id}
+        )`.as('roleCount'),
+      },
+      with: {
+        roles: {
+          with: {
+            role: true,
+          },
+        },
+      },
+      columns: {
+        isDeleted: false,
+        deletedAt: false,
+        deletedBy: false,
+        deletedReason: false,
+      },
     });
 
     if (!item) {
       throw new NotFoundException('Permission not found');
     }
 
-    return item;
+    return {
+      message: 'Permission fetched successfully',
+      item: item,
+    };
   }
 
   async update(id: number, updatePermissionDto: UpdatePermissionDto) {
@@ -173,6 +273,9 @@ export class PermissionsService {
         name: updatePermissionDto.name ? updatePermissionDto.name : data.name,
         slug: slugify(updatePermissionDto.name || ''),
         group: updatePermissionDto.group ? updatePermissionDto.group : data.group,
+        groupOrder: updatePermissionDto.groupOrder ? updatePermissionDto.groupOrder : data.groupOrder,
+        order: updatePermissionDto.order ? updatePermissionDto.order : data.order,
+        isActive: updatePermissionDto.isActive ? updatePermissionDto.isActive : data.isActive,
         updatedAt: new Date(),
       })
       .where(eq(permissions.id, id));
@@ -211,6 +314,74 @@ export class PermissionsService {
       throw new HttpException('Permission has roles, cannot be deleted', HttpStatus.BAD_REQUEST);
     }
 
+    await this.db
+      .update(permissions)
+      .set({
+        isDeleted: true,
+        deletedAt: new Date(),
+        deletedBy: 1,
+        deletedReason: 'Deleted by admin',
+      })
+      .where(eq(permissions.id, id));
+
+    return {
+      message: 'Permission deleted successfully',
+    };
+  }
+
+  async restore(id: number) {
+    const item = await this.db.query.permissions.findFirst({
+      where: eq(permissions.id, id),
+    });
+
+    if (!item) {
+      throw new NotFoundException('Permission not found');
+    }
+
+    await this.db
+      .update(permissions)
+      .set({
+        isDeleted: false,
+        deletedAt: null,
+        deletedBy: null,
+        deletedReason: null,
+      })
+      .where(eq(permissions.id, id));
+
+    return {
+      message: 'Permission restored successfully',
+    };
+  }
+
+  async forceDelete(id: number) {
+    const item = await this.db.query.permissions.findFirst({
+      where: eq(permissions.id, id),
+      extras: {
+        adminCount: sql<number>`(
+          SELECT COUNT(*)::int 
+          FROM ${adminPermission} ap
+          WHERE ap.permission_id = ${permissions.id}
+        )`.as('adminCount'),
+        roleCount: sql<number>`(
+          SELECT COUNT(*)::int 
+          FROM ${permissionRole} pr
+          WHERE pr.permission_id = ${permissions.id}
+        )`.as('roleCount'),
+      },
+    });
+
+    if (!item) {
+      throw new NotFoundException('Permission not found');
+    }
+
+    if (item.adminCount > 0) {
+      throw new HttpException('Permission has admins, cannot be deleted', HttpStatus.BAD_REQUEST);
+    }
+
+    if (item.roleCount > 0) {
+      throw new HttpException('Permission has roles, cannot be deleted', HttpStatus.BAD_REQUEST);
+    }
+
     await this.db.delete(permissions).where(eq(permissions.id, id));
 
     return {
@@ -220,11 +391,67 @@ export class PermissionsService {
 
   // getAllPermissions
   async getAllPermissions() {
-    const items = await this.db.select().from(permissions).orderBy(asc(permissions.slug));
+    const items = await this.db
+      .select({
+        id: permissions.id,
+        name: permissions.name,
+        slug: permissions.slug,
+        group: permissions.group,
+        groupOrder: permissions.groupOrder,
+        order: permissions.order,
+        isActive: permissions.isActive,
+        createdAt: permissions.createdAt,
+        updatedAt: permissions.updatedAt,
+      })
+      .from(permissions)
+      .where(and(eq(permissions.isDeleted, false), eq(permissions.isActive, true)))
+      .orderBy(asc(permissions.id));
 
     return {
       message: 'Items fetched successfully',
       items: items,
+    };
+  }
+
+  async changeStatus(id: number) {
+    const item = await this.db.query.permissions.findFirst({
+      where: eq(permissions.id, id),
+      extras: {
+        adminCount: sql<number>`(
+          SELECT COUNT(*)::int 
+          FROM ${adminPermission} ap
+          WHERE ap.permission_id = ${permissions.id}
+        )`.as('adminCount'),
+        roleCount: sql<number>`(
+          SELECT COUNT(*)::int 
+          FROM ${permissionRole} pr
+          WHERE pr.permission_id = ${permissions.id}
+        )`.as('roleCount'),
+      },
+    });
+
+    if (!item) {
+      throw new HttpException('Permission not found', HttpStatus.BAD_REQUEST);
+    }
+
+    if (item.adminCount > 0) {
+      throw new HttpException('Permission has admins, cannot change status', HttpStatus.BAD_REQUEST);
+    }
+
+    if (item.roleCount > 0) {
+      throw new HttpException('Permission has roles, cannot change status', HttpStatus.BAD_REQUEST);
+    }
+
+    await this.db
+      .update(permissions)
+      .set({
+        isActive: !item.isActive,
+        updatedAt: new Date(),
+      })
+      .where(eq(permissions.id, id));
+
+    return {
+      message: 'Status Changed successfully',
     };
   }
 }
